Load backends automatically if they exist in the backends package
Register click commands located in specific backend modules.
So in order to add a new backend one has to implement a click command
which utilizs a `session` and `render` function passed in the command
context. Note that the command has to have the same name as the
specific backend module.